awesome rc.lua update,start.sh fixed
due to the version bump of awesome wm, original rc.lua dosen't work any
more, so copy new rc.lua from /etc/xdg/awesome/ and modified to let the
control bar at the bottom of screen, tiling layout to be the default
one.

Another thing, when first version of start.sh was finished, every time
we ps aux and then grep by process name,such as "ps aux|grep firefox",
at least one line exist in the grep results,that is the "grep firefox"
command itself. But at the day of 20170411, I found that behavior of
changed,  in the results of ps |grep command, "grep command" will not show up.
so the logical to check whether a program is running or not need to be
changed to fit this new behavior of "ps aux|grep".
wc -l == "1" changed to wc -l =="0"